+++ src/components/modals/index.ts
@@ -1,0 +1,2 @@
+export { CreateProjectModal } from "./create-project-modal";
+