+++ src/components/views/index.ts
@@ -1,0 +1,2 @@
+export { MindMapView } from "./mind-map-view";
+